+++ src/apps/server/experiences/controllers/experience.controller.ts
@@ -16,6 +16,12 @@ import { ResponseEntity } from '📚libs/utils/respone.entity';
 import { GetExperienceNotFoundErrorResDto, GetExperienceResDto } from '../dto/res/getExperience.res.dto';
 import { Method } from '📚libs/enums/method.enum';
 import { getExperienceSuccMd, upsertExperienceSuccMd } from '🔥apps/server/experiences/markdown/experience.md';
+import {
+  GetCountOfExperienceAndCapabilityDescriptionMd,
+  GetCountOfExperienceAndCapabilityResponseDescriptionMd,
+  GetCountOfExperienceAndCapabilitySummaryMd,
+} from '../markdown/get-count-of-experience-and-capability.doc';
+import { GetCountOfExperienceAndCapabilityResponseDto } from '🔥apps/server/experiences/dto/get-count-of-experience-and-capability.dto';
 
 @ApiBearerAuth()
 @UseGuards(JwtAuthGuard)
@@ -74,5 +80,27 @@ export class ExperienceController {
     const experience = await this.experienceService.getExperienceByUserId(user.userId);
 
     return ResponseEntity.OK_WITH_DATA(experience);
+  }
+
+  @Route({
+    request: {
+      path: '/count',
+      method: Method.GET,
+    },
+    response: {
+      code: HttpStatus.OK,
+      type: GetCountOfExperienceAndCapabilityResponseDto,
+      isArray: true,
+      description: GetCountOfExperienceAndCapabilityResponseDescriptionMd,
+    },
+    summary: GetCountOfExperienceAndCapabilitySummaryMd,
+    description: GetCountOfExperienceAndCapabilityDescriptionMd,
+  })
+  async getCountOfExperienceAndCapability(
+    @User() user: UserJwtToken,
+  ): Promise<ResponseEntity<GetCountOfExperienceAndCapabilityResponseDto[]>> {
+    const countOfExperienceAndCapability = await this.experienceService.getCountOfExperienceAndCapability(user.userId);
+
+    return ResponseEntity.OK_WITH_DATA(countOfExperienceAndCapability);
   }
 }
