+++ src/apps/server/experiences/provider/experience.provider.ts
@@ -8,4 +8,5 @@ export const experienceProviders = [
   ExperienceInfoRepository,
   CapabilityRepository,
   ExperienceCapabilityRepository,
+  CapabilityRepository,
 ] as const;
