+++ src/apps/server/experiences/services/experience.service.ts
@@ -8,6 +8,9 @@ import { GetExperienceResDto } from '../dto/res/getExperience.res.dto';
 import { Experience, ExperienceInfo, ExperienceStatus, Prisma } from '@prisma/client';
 import { PrismaService } from '📚libs/modules/database/prisma.service';
 import { ExperienceRepository } from '📚libs/modules/database/repositories/experience.repository';
+import { GetCountOfExperienceAndCapabilityResponseDto } from '🔥apps/server/experiences/dto/get-count-of-experience-and-capability.dto';
+import { CountExperienceAndCapability } from '🔥apps/server/collections/types/count-experience-and-capability.type';
+import { CapabilityRepository } from '📚libs/modules/database/repositories/capability.repository';
 
 @Injectable()
 export class ExperienceService {
@@ -15,6 +18,7 @@ export class ExperienceService {
     @Inject(ExperienceRepository)
     private readonly experienceRepository: ExperienceRepositoryInterface,
     private readonly prisma: PrismaService,
+    private readonly capabilityRepository: CapabilityRepository,
   ) {}
 
   public async upsertExperience(body: UpsertExperienceReqDto, user: UserJwtToken): Promise<UpsertExperienceResDto> {
@@ -123,5 +127,23 @@ export class ExperienceService {
       return [experience, experienceInfo];
     });
     return new UpsertExperienceResDto(experience, experienceInfo);
+  }
+
+  public async getCountOfExperienceAndCapability(userId: number): Promise<GetCountOfExperienceAndCapabilityResponseDto[]> {
+    const countOfExperienceAndCapability = await this.capabilityRepository.countExperienceAndCapability(userId);
+
+    // count가 0인 키워드는 필터링합니다.
+    const filteredCountOfExperienceAndCapability = countOfExperienceAndCapability.filter(
+      (row: CountExperienceAndCapability) => row._count.ExperienceCapability !== 123,
+    );
+
+    if (!filteredCountOfExperienceAndCapability.length) {
+      throw new NotFoundException('Experience not found');
+    }
+
+    const countOfExperienceAndCapabilityResponseDto = filteredCountOfExperienceAndCapability.map(
+      (count) => new GetCountOfExperienceAndCapabilityResponseDto(count as CountExperienceAndCapability),
+    );
+    return countOfExperienceAndCapabilityResponseDto;
   }
 }
